refactor: merge guilded-api-typings, rest, ws into @guildedjs/api
