+++ .github/CODEOWNER
@@ -1,4 +1,2 @@
-/packages/rest @Skillz4Killz @zaida04
-/packages/ws @Skillz4Killz @zaida04
-/packages/guilded.js @Skillz4Killz @zaida04
-/packages/guilded-api-types @zaida04
+/packages/api @zaida04
+/packages/guilded.js @zaida04
