+++ packages/api/README.md
@@ -1,19 +1,19 @@
-# `@guildedjs/rest`
+# `@guildedjs/api`
 
 [![GitHub](https://img.shields.io/badge/License-MIT-yellow.svg)](https://github.com/zaida04/guilded.js/blob/main/LICENSE)
-[![npm](https://img.shields.io/npm/v/@guildedjs/rest?color=crimson&logo=npm)](https://www.npmjs.com/package/@guildedjs/rest)
+[![npm](https://img.shields.io/npm/v/@guildedjs/api?color=crimson&logo=npm)](https://www.npmjs.com/package/@guildedjs/api)
 [![CI workflows](https://github.com/zaida04/guilded.js/actions/workflows/ci.yml/badge.svg)](https://github.com/zaida04/guilded.js/actions/workflows/ci.yml)
 
-Rest utility for making requests to the Guilded.gg API.
+Multi-use wrapper over the entire Guilded APi. Utilities for making REST requests to the Guilded API with ratelimit handling. Also contains utilities for connecting to Guilded's WebSocket gateway and receiving events.
 
 > ⚠️ If you intend to use `guilded.js` you won't need to install this, as it's already a dependency of that package.
 
 ## Installation
 
-You can install this package from [NPM](https://www.npmjs.com/package/@guildedjs/rest)
+You can install this package from [NPM](https://www.npmjs.com/package/@guildedjs/api)
 
-- `npm install @guildedjs/rest`
-- `yarn add @guildedjs/rest`
+- `npm install @guildedjs/api`
+- `yarn add @guildedjs/api`
 
 ## Contributing
