+++ packages/api/lib/api-typings/v1/structs/Channel.ts
@@ -1,0 +1,11 @@
+export type ChannelType =
+  | "announcements"
+  | "calendar"
+  | "chat"
+  | "docs"
+  | "forums"
+  | "list"
+  | "media"
+  | "scheduling"
+  | "stream"
+  | "voice";
