+++ packages/api/lib/api-typings/v1/ws.ts
@@ -3,32 +3,35 @@
  * Do not make direct changes to the file.
  */
 
-
 export type paths = Record<string, never>;
 
 export type webhooks = Record<string, never>;
 
-export interface components {
+export type components = {
   schemas: {
     _SocketEventEnvelope: {
       /**
-       * Opcode 
-       * @description An operation code corresponding to the nature of the sent message (for example, success, failure, etc.) 
+       * Opcode
+       *
+       * @description An operation code corresponding to the nature of the sent message (for example, success, failure, etc.)
        * @enum {integer}
        */
       op: 0 | 1 | 2 | 8 | 9;
       /**
-       * Data 
+       * Data
+       *
        * @description Data of any form depending on the underlying event
        */
       d?: Record<string, never>;
       /**
-       * Message ID 
+       * Message ID
+       *
        * @description Message ID used for replaying events after a disconnect
        */
       s?: string;
       /**
-       * Event name 
+       * Event name
+       *
        * @description Event name for the given message
        */
       t?: string;
@@ -87,73 +90,85 @@ export interface components {
      */
     ChatMessage: {
       /**
-       * Message ID 
-       * Format: uuid 
+       * Message ID
+       * Format: uuid
+       *
        * @description The ID of the message
        */
       id: string;
       /**
-       * Type 
-       * @description The type of chat message. "system" messages are generated by Guilded, while "default" messages are user or bot-generated. 
+       * Type
+       *
+       * @description The type of chat message. "system" messages are generated by Guilded, while "default" messages are user or bot-generated.
        * @enum {string}
        */
       type: "default" | "system";
       /**
-       * Server ID 
+       * Server ID
+       *
        * @description The ID of the server
        */
       serverId?: string;
       /**
-       * Group ID 
+       * Group ID
+       *
        * @description The ID of the group
        */
       groupId?: string;
       /**
-       * Channel ID 
-       * Format: uuid 
+       * Channel ID
+       * Format: uuid
+       *
        * @description The ID of the channel
        */
       channelId: string;
       /**
-       * Content 
-       * Format: markdown 
+       * Content
+       * Format: markdown
+       *
        * @description The content of the message
        */
       content?: string;
-      embeds?: (components["schemas"]["ChatEmbed"])[];
+      embeds?: components["schemas"]["ChatEmbed"][];
       /** @description Message IDs that were replied to */
-      replyMessageIds?: (string)[];
+      replyMessageIds?: string[];
       /**
-       * Is private 
+       * Is private
+       *
        * @description If set, this message will only be seen by those mentioned or replied to
        */
       isPrivate?: boolean;
       /**
-       * Is silent 
-       * @description If set, this message did not notify mention or reply recipients 
+       * Is silent
+       *
+       * @description If set, this message did not notify mention or reply recipients
        * @default false
        */
       isSilent?: boolean;
       mentions?: components["schemas"]["Mentions"];
       /**
-       * Created at 
-       * Format: date-time 
+       * Created at
+       * Format: date-time
+       *
        * @description The ISO 8601 timestamp that the message was created at
        */
       createdAt: string;
       /**
-       * Created by 
+       * Created by
+       *
        * @description The ID of the user who created this message (Note: If this event has `createdByWebhookId` present, this field will still be populated, but can be ignored. In this case, the value of this field will always be Ann6LewA)
        */
       createdBy: string;
       /**
-       * Created by Webhook ID 
+       * Created by Webhook ID
+       *
        * @description The ID of the webhook who created this message, if it was created by a webhook
        */
       createdByWebhookId?: string;
       /**
-       * Updated at 
-       * Format: date-time 
+       * Updated at
+       * Format: date-time
+       *
        * @description The ISO 8601 timestamp that the message was updated at, if relevant
        */
       updatedAt?: string;
@@ -172,26 +187,29 @@ export interface components {
      */
     ChatMessageReaction: {
       /**
-       * Channel ID 
-       * Format: uuid 
+       * Channel ID
+       * Format: uuid
+       *
        * @description The ID of the channel
        */
       channelId: string;
       /**
-       * User ID 
+       * User ID
+       *
        * @description The ID of the user who added the reaction
        */
       createdBy: string;
       emote: components["schemas"]["Emote"];
       /**
-       * Message ID 
-       * Format: uuid 
+       * Message ID
+       * Format: uuid
+       *
        * @description The ID of the message
        */
       messageId: string;
     };
     /**
-     * @description Rich content sections optionally associated with chat messages. Properties with "webhook-markdown" support allow for the following: link, italic, bold, strikethrough, underline, inline code, block code, reaction, and mention. 
+     * @description Rich content sections optionally associated with chat messages. Properties with "webhook-markdown" support allow for the following: link, italic, bold, strikethrough, underline, inline code, block code, reaction, and mention.
      * @example {
      *   "title": "embed title",
      *   "description": "embeds support a **different** __subset__ *of* markdown than other markdown fields. <@Ann6LewA>\n\n [links](https://www.guilded.gg) ```\ncheck this code out```\n\n:pizza: time!! ttyl",
@@ -233,17 +251,20 @@ export interface components {
      */
     ChatEmbed: {
       /**
-       * Format: webhook-markdown 
+       * Format: webhook-markdown
+       *
        * @description Main header of the embed
        */
       title?: string;
       /**
-       * Format: webhook-markdown 
+       * Format: webhook-markdown
+       *
        * @description Subtext of the embed
        */
       description?: string;
       /**
-       * Format: uri 
+       * Format: uri
+       *
        * @description URL to linkify the `title` field with
        */
       url?: string;
@@ -252,7 +273,8 @@ export interface components {
       /** @description A small section at the bottom of the embed */
       footer?: {
         /**
-         * Format: media-uri 
+         * Format: media-uri
+         *
          * @description URL of a small image to put in the footer
          */
         icon_url?: string;
@@ -260,14 +282,16 @@ export interface components {
         text: string;
       };
       /**
-       * Format: date-time 
+       * Format: date-time
+       *
        * @description A timestamp to put in the footer
        */
       timestamp?: string;
       /** @description An image to the right of the embed's content */
       thumbnail?: {
         /**
-         * Format: media-uri 
+         * Format: media-uri
+         *
          * @description URL of the image
          */
         url?: string;
@@ -275,7 +299,8 @@ export interface components {
       /** @description The main picture to associate with the embed */
       image?: {
         /**
-         * Format: media-uri 
+         * Format: media-uri
+         *
          * @description URL of the image
          */
         url?: string;
@@ -285,34 +310,38 @@ export interface components {
         /** @description Name of the author */
         name?: string;
         /**
-         * Format: uri 
+         * Format: uri
+         *
          * @description URL to linkify the author's `name` field
          */
         url?: string;
         /**
-         * Format: media-uri 
+         * Format: media-uri
+         *
          * @description URL of a small image to display to the left of the author's `name`
          */
         icon_url?: string;
       };
       /** @description Table-like cells to add to the embed */
-      fields?: ({
-          /**
-           * Format: webhook-markdown 
-           * @description Header of the table-like cell
-           */
-          name: string;
-          /**
-           * Format: webhook-markdown 
-           * @description Subtext of the table-like cell
-           */
-          value: string;
-          /**
-           * @description If the field should wrap or not 
-           * @default false
-           */
-          inline?: boolean;
-        })[];
+      fields?: {
+        /**
+         * Format: webhook-markdown
+         *
+         * @description Header of the table-like cell
+         */
+        name: string;
+        /**
+         * Format: webhook-markdown
+         *
+         * @description Subtext of the table-like cell
+         */
+        value: string;
+        /**
+         * @description If the field should wrap or not
+         * @default false
+         */
+        inline?: boolean;
+      }[];
     };
     /**
      * @example {
@@ -325,13 +354,16 @@ export interface components {
      */
     SocialLink: {
       /**
-       * Social link type 
-       * @description The type of social link that Guilded supports. Depending on this value, `handle` or `serviceId` may or may not be present 
+       * Social link type
+       *
+       * @description The type of social link that Guilded supports. Depending on this value, `handle` or `serviceId` may or may not be present
        * @enum {string}
        */
-      type: "twitch" | "bnet" | "psn" | "xbox" | "steam" | "origin" | "youtube" | "twitter" | "facebook" | "switch" | "patreon" | "roblox" | "epic";
+      type:
+        "bnet" | "epic" | "facebook" | "origin" | "patreon" | "psn" | "roblox" | "steam" | "switch" | "twitch" | "twitter" | "xbox" | "youtube";
       /**
-       * User ID 
+       * User ID
+       *
        * @description The ID of the user that the social link is associated with
        */
       userId: string;
@@ -340,14 +372,15 @@ export interface components {
       /** @description The unique ID that represents this member's social link within the external service */
       serviceId?: string;
       /**
-       * Created at 
-       * Format: date-time 
+       * Created at
+       * Format: date-time
+       *
        * @description The ISO 8601 timestamp that the social link was created at
        */
       createdAt: string;
     };
     /**
-     * @description Metadata of who or what is mentioned in content 
+     * @description Metadata of who or what is mentioned in content
      * @example {
      *   "users": [
      *     {
@@ -370,46 +403,54 @@ export interface components {
      */
     Mentions: {
       /**
-       * Users 
+       * Users
+       *
        * @description Info on mentioned users
        */
-      users?: ({
-          /**
-           * User ID 
-           * @description The ID of the user
-           */
-          id: string;
-        })[];
+      users?: {
+        /**
+         * User ID
+         *
+         * @description The ID of the user
+         */
+        id: string;
+      }[];
       /**
-       * Channels 
+       * Channels
+       *
        * @description Info on mentioned channels
        */
-      channels?: ({
-          /**
-           * Channel ID 
-           * Format: uuid 
-           * @description The ID of the channel
-           */
-          id: string;
-        })[];
+      channels?: {
+        /**
+         * Channel ID
+         * Format: uuid
+         *
+         * @description The ID of the channel
+         */
+        id: string;
+      }[];
       /**
-       * Roles 
+       * Roles
+       *
        * @description Info on mentioned roles
        */
-      roles?: ({
-          /**
-           * Role ID 
-           * @description The ID of the role
-           */
-          id: number;
-        })[];
+      roles?: {
+        /**
+         * Role ID
+         *
+         * @description The ID of the role
+         */
+        id: number;
+      }[];
       /**
-       * Everyone 
+       * Everyone
+       *
        * @description If @everyone was mentioned
        */
       everyone?: boolean;
       /**
-       * Here 
+       * Here
+       *
        * @description If @here was mentioned
        */
       here?: boolean;
@@ -428,33 +469,38 @@ export interface components {
       /** @description The ID of the forum topic comment */
       id: number;
       /**
-       * Content 
-       * Format: markdown 
+       * Content
+       * Format: markdown
+       *
        * @description The content of the forum topic comment
        */
       content: string;
       /**
-       * Created at 
-       * Format: date-time 
+       * Created at
+       * Format: date-time
+       *
        * @description The ISO 8601 timestamp that the forum topic comment was created at
        */
       createdAt: string;
       /**
-       * Updated at 
-       * Format: date-time 
+       * Updated at
+       * Format: date-time
+       *
        * @description The ISO 8601 timestamp that the forum topic comment was updated at, if relevant
        */
       updatedAt?: string;
       /**
-       * Channel ID 
-       * Format: uuid 
+       * Channel ID
+       * Format: uuid
+       *
        * @description The ID of the channel
        */
       channelId: string;
       /** @description The ID of the forum topic */
       forumTopicId: number;
       /**
-       * Created by 
+       * Created by
+       *
        * @description The ID of the user who created this forum topic comment (Note: If this event has `createdByWebhookId` present, this field will still be populated, but can be ignored. In this case, the value of this field will always be Ann6LewA)
        */
       createdBy: string;
@@ -474,13 +520,15 @@ export interface components {
      */
     ForumTopicReaction: {
       /**
-       * Channel ID 
-       * Format: uuid 
+       * Channel ID
+       * Format: uuid
+       *
        * @description The ID of the channel
        */
       channelId: string;
       /**
-       * User ID 
+       * User ID
+       *
        * @description The ID of the user who added the reaction
        */
       createdBy: string;
@@ -503,13 +551,15 @@ export interface components {
      */
     ForumTopicCommentReaction: {
       /**
-       * Channel ID 
-       * Format: uuid 
+       * Channel ID
+       * Format: uuid
+       *
        * @description The ID of the channel
        */
       channelId: string;
       /**
-       * User ID 
+       * User ID
+       *
        * @description The ID of the user who added the reaction
        */
       createdBy: string;
@@ -534,57 +584,67 @@ export interface components {
       /** @description The ID of the forum topic */
       id: number;
       /**
-       * Server ID 
+       * Server ID
+       *
        * @description The ID of the server
        */
       serverId: string;
       /**
-       * Channel ID 
-       * Format: uuid 
+       * Channel ID
+       * Format: uuid
+       *
        * @description The ID of the channel
        */
       channelId: string;
       /**
-       * Title 
+       * Title
+       *
        * @description The title of the forum topic
        */
       title: string;
       /**
-       * Created at 
-       * Format: date-time 
+       * Created at
+       * Format: date-time
+       *
        * @description The ISO 8601 timestamp that the forum topic was created at
        */
       createdAt: string;
       /**
-       * Created by 
+       * Created by
+       *
        * @description The ID of the user who created this forum topic (Note: If this event has `createdByWebhookId` present, this field will still be populated, but can be ignored. In this case, the value of this field will always be Ann6LewA)
        */
       createdBy: string;
       /**
-       * Updated at 
-       * Format: date-time 
+       * Updated at
+       * Format: date-time
+       *
        * @description The ISO 8601 timestamp that the forum topic was updated at, if relevant
        */
       updatedAt?: string;
       /**
-       * Bumped at 
-       * Format: date-time 
+       * Bumped at
+       * Format: date-time
+       *
        * @description The ISO 8601 timestamp that the forum topic was bumped at. This timestamp is updated whenever there is any activity on the posts within the forum topic.
        */
       bumpedAt?: string;
       /**
-       * Is pinned 
+       * Is pinned
+       *
        * @default false
        */
       isPinned?: boolean;
       /**
-       * Is locked 
+       * Is locked
+       *
        * @default false
        */
       isLocked?: boolean;
       /**
-       * Content 
-       * Format: markdown 
+       * Content
+       * Format: markdown
+       *
        * @description The content of the forum topic
        */
       content: string;
@@ -604,51 +664,60 @@ export interface components {
       /** @description The ID of the forum topic */
       id: number;
       /**
-       * Server ID 
+       * Server ID
+       *
        * @description The ID of the server
        */
       serverId: string;
       /**
-       * Channel ID 
-       * Format: uuid 
+       * Channel ID
+       * Format: uuid
+       *
        * @description The ID of the channel
        */
       channelId: string;
       /**
-       * Title 
+       * Title
+       *
        * @description The title of the forum topic
        */
       title: string;
       /**
-       * Created at 
-       * Format: date-time 
+       * Created at
+       * Format: date-time
+       *
        * @description The ISO 8601 timestamp that the forum topic was created at
        */
       createdAt: string;
       /**
-       * Created by 
+       * Created by
+       *
        * @description The ID of the user who created this forum topic (Note: If this event has `createdByWebhookId` present, this field will still be populated, but can be ignored. In this case, the value of this field will always be Ann6LewA)
        */
       createdBy: string;
       /**
-       * Updated at 
-       * Format: date-time 
+       * Updated at
+       * Format: date-time
+       *
        * @description The ISO 8601 timestamp that the forum topic was updated at, if relevant
        */
       updatedAt?: string;
       /**
-       * Bumped at 
-       * Format: date-time 
+       * Bumped at
+       * Format: date-time
+       *
        * @description The ISO 8601 timestamp that the forum topic was bumped at. This timestamp is updated whenever there is any activity on the posts within the forum topic.
        */
       bumpedAt?: string;
       /**
-       * Is pinned 
+       * Is pinned
+       *
        * @default false
        */
       isPinned?: boolean;
       /**
-       * Is locked 
+       * Is locked
+       *
        * @default false
        */
       isLocked?: boolean;
@@ -670,98 +739,115 @@ export interface components {
      */
     ListItem: {
       /**
-       * Format: uuid 
+       * Format: uuid
+       *
        * @description The ID of the list item
        */
       id: string;
       /**
-       * Server ID 
+       * Server ID
+       *
        * @description The ID of the server
        */
       serverId: string;
       /**
-       * Channel ID 
-       * Format: uuid 
+       * Channel ID
+       * Format: uuid
+       *
        * @description The ID of the channel
        */
       channelId: string;
       /**
-       * Message 
-       * Format: markdown 
+       * Message
+       * Format: markdown
+       *
        * @description The message of the list item
        */
       message: string;
       mentions?: components["schemas"]["Mentions"];
       /**
-       * Created at 
-       * Format: date-time 
+       * Created at
+       * Format: date-time
+       *
        * @description The ISO 8601 timestamp that the list item was created at
        */
       createdAt: string;
       /**
-       * Created by 
+       * Created by
+       *
        * @description The ID of the user who created this list item (Note: If this event has `createdByWebhookId` present, this field will still be populated, but can be ignored. In this case, the value of this field will always be Ann6LewA)
        */
       createdBy: string;
       /**
-       * Created by Webhook ID 
+       * Created by Webhook ID
+       *
        * @description The ID of the webhook who created this list item, if it was created by a webhook
        */
       createdByWebhookId?: string;
       /**
-       * Updated at 
-       * Format: date-time 
+       * Updated at
+       * Format: date-time
+       *
        * @description The ISO 8601 timestamp that the list item was updated at, if relevant
        */
       updatedAt?: string;
       /**
-       * Updated by 
+       * Updated by
+       *
        * @description The ID of the user who updated this list item
        */
       updatedBy?: string;
       /**
-       * Format: uuid 
+       * Format: uuid
+       *
        * @description The ID of the parent list item if this list item is nested
        */
       parentListItemId?: string;
       /**
-       * Completed at 
-       * Format: date-time 
+       * Completed at
+       * Format: date-time
+       *
        * @description The ISO 8601 timestamp that the list item was completed at
        */
       completedAt?: string;
       /**
-       * Completed by 
+       * Completed by
+       *
        * @description The ID of the user who completed this list item
        */
       completedBy?: string;
       note?: {
         /**
-         * Created at 
-         * Format: date-time 
+         * Created at
+         * Format: date-time
+         *
          * @description The ISO 8601 timestamp that the note was created at. If this field is populated, then there's a note associated with the list item
          */
         createdAt: string;
         /**
-         * Created by 
+         * Created by
+         *
          * @description The ID of the user who created this note
          */
         createdBy: string;
         /**
-         * Updated at 
-         * Format: date-time 
+         * Updated at
+         * Format: date-time
+         *
          * @description The ISO 8601 timestamp that the note was updated at, if relevant
          */
         updatedAt?: string;
         /**
-         * Updated by 
+         * Updated by
+         *
          * @description The ID of the user who updated this note
          */
         updatedBy?: string;
         mentions?: components["schemas"]["Mentions"];
         /**
-         * Note 
-         * Format: markdown 
+         * Note
+         * Format: markdown
+         *
          * @description The note of the list item
          */
         content: string;
@@ -783,91 +869,107 @@ export interface components {
      */
     ListItemSummary: {
       /**
-       * Format: uuid 
+       * Format: uuid
+       *
        * @description The ID of the list item
        */
       id: string;
       /**
-       * Server ID 
+       * Server ID
+       *
        * @description The ID of the server
        */
       serverId: string;
       /**
-       * Channel ID 
-       * Format: uuid 
+       * Channel ID
+       * Format: uuid
+       *
        * @description The ID of the channel
        */
       channelId: string;
       /**
-       * Message 
-       * Format: markdown 
+       * Message
+       * Format: markdown
+       *
        * @description The message of the list item
        */
       message: string;
       mentions?: components["schemas"]["Mentions"];
       /**
-       * Created at 
-       * Format: date-time 
+       * Created at
+       * Format: date-time
+       *
        * @description The ISO 8601 timestamp that the list item was created at
        */
       createdAt: string;
       /**
-       * Created by 
+       * Created by
+       *
        * @description The ID of the user who created this list item (Note: If this event has `createdByWebhookId` present, this field will still be populated, but can be ignored. In this case, the value of this field will always be Ann6LewA)
        */
       createdBy: string;
       /**
-       * Created by Webhook ID 
+       * Created by Webhook ID
+       *
        * @description The ID of the webhook who created this list item, if it was created by a webhook
        */
       createdByWebhookId?: string;
       /**
-       * Updated at 
-       * Format: date-time 
+       * Updated at
+       * Format: date-time
+       *
        * @description The ISO 8601 timestamp that the list item was updated at, if relevant
        */
       updatedAt?: string;
       /**
-       * Updated by 
+       * Updated by
+       *
        * @description The ID of the user who updated this list item
        */
       updatedBy?: string;
       /**
-       * Format: uuid 
+       * Format: uuid
+       *
        * @description The ID of the parent list item if this list item is nested
        */
       parentListItemId?: string;
       /**
-       * Completed at 
-       * Format: date-time 
+       * Completed at
+       * Format: date-time
+       *
        * @description The ISO 8601 timestamp that the list item was completed at
        */
       completedAt?: string;
       /**
-       * Completed by 
+       * Completed by
+       *
        * @description The ID of the user who completed this list item
        */
       completedBy?: string;
       note?: {
         /**
-         * Created at 
-         * Format: date-time 
+         * Created at
+         * Format: date-time
+         *
          * @description The ISO 8601 timestamp that the note was created at. If this field is populated, then there's a note associated with the list item
          */
         createdAt: string;
         /**
-         * Created by 
+         * Created by
+         *
          * @description The ID of the user who created this note
          */
         createdBy: string;
         /**
-         * Updated at 
-         * Format: date-time 
+         * Updated at
+         * Format: date-time
+         *
          * @description The ISO 8601 timestamp that the note was updated at, if relevant
          */
         updatedAt?: string;
         /**
-         * Updated by 
+         * Updated by
+         *
          * @description The ID of the user who updated this note
          */
         updatedBy?: string;
@@ -888,52 +990,61 @@ export interface components {
      */
     Doc: {
       /**
-       * Doc ID 
+       * Doc ID
+       *
        * @description The ID of the doc
        */
       id: number;
       /**
-       * Server ID 
+       * Server ID
+       *
        * @description The ID of the server
        */
       serverId: string;
       /**
-       * Channel ID 
-       * Format: uuid 
+       * Channel ID
+       * Format: uuid
+       *
        * @description The ID of the channel
        */
       channelId: string;
       /**
-       * Title 
+       * Title
+       *
        * @description The title of the doc
        */
       title: string;
       /**
-       * Content 
-       * Format: markdown 
+       * Content
+       * Format: markdown
+       *
        * @description The content of the doc
        */
       content: string;
       mentions?: components["schemas"]["Mentions"];
       /**
-       * Created at 
-       * Format: date-time 
+       * Created at
+       * Format: date-time
+       *
        * @description The ISO 8601 timestamp that the doc was created at
        */
       createdAt: string;
       /**
-       * Created by 
+       * Created by
+       *
        * @description The ID of the user who created this doc
        */
       createdBy: string;
       /**
-       * Updated at 
-       * Format: date-time 
+       * Updated at
+       * Format: date-time
+       *
        * @description The ISO 8601 timestamp that the doc was updated at, if relevant
        */
       updatedAt?: string;
       /**
-       * Updated by 
+       * Updated by
+       *
        * @description The ID of the user who updated this doc
        */
       updatedBy?: string;
@@ -950,41 +1061,48 @@ export interface components {
      */
     DocComment: {
       /**
-       * Doc comment ID 
+       * Doc comment ID
+       *
        * @description The ID of the doc comment
        */
       id: number;
       /**
-       * Content 
-       * Format: markdown 
+       * Content
+       * Format: markdown
+       *
        * @description The content of the doc comment
        */
       content: string;
       /**
-       * Created at 
-       * Format: date-time 
+       * Created at
+       * Format: date-time
+       *
        * @description The ISO 8601 timestamp that the doc comment was created at
        */
       createdAt: string;
       /**
-       * Created by 
+       * Created by
+       *
        * @description The ID of the user who created this doc comment (Note: If this event has `createdByWebhookId` present, this field will still be populated, but can be ignored. In this case, the value of this field will always be Ann6LewA)
        */
       createdBy: string;
       /**
-       * Updated at 
-       * Format: date-time 
+       * Updated at
+       * Format: date-time
+       *
        * @description The ISO 8601 timestamp that the doc comment was updated at, if relevant
        */
       updatedAt?: string;
       /**
-       * Channel ID 
-       * Format: uuid 
+       * Channel ID
+       * Format: uuid
+       *
        * @description The ID of the channel
        */
       channelId: string;
       /**
-       * Doc ID 
+       * Doc ID
+       *
        * @description The ID of the doc
        */
       docId: number;
@@ -1004,19 +1122,22 @@ export interface components {
      */
     DocReaction: {
       /**
-       * Channel ID 
-       * Format: uuid 
+       * Channel ID
+       * Format: uuid
+       *
        * @description The ID of the channel
        */
       channelId: string;
       /**
-       * User ID 
+       * User ID
+       *
        * @description The ID of the user who added the reaction
        */
       createdBy: string;
       emote: components["schemas"]["Emote"];
       /**
-       * Doc ID 
+       * Doc ID
+       *
        * @description The ID of the doc
        */
       docId: number;
@@ -1036,24 +1157,28 @@ export interface components {
      */
     DocCommentReaction: {
       /**
-       * Channel ID 
-       * Format: uuid 
+       * Channel ID
+       * Format: uuid
+       *
        * @description The ID of the channel
        */
       channelId: string;
       /**
-       * User ID 
+       * User ID
+       *
        * @description The ID of the user who added the reaction
        */
       createdBy: string;
       emote: components["schemas"]["Emote"];
       /**
-       * Doc ID 
+       * Doc ID
+       *
        * @description The ID of the doc
        */
       docId: number;
       /**
-       * Doc comment ID 
+       * Doc comment ID
+       *
        * @description The ID of the doc comment
        */
       docCommentId: number;
@@ -1074,17 +1199,19 @@ export interface components {
     ServerMember: {
       user: components["schemas"]["User"];
       /** Role IDs */
-      roleIds: (number)[];
+      roleIds: number[];
       /** Nickname */
       nickname?: string;
       /**
-       * Created at 
-       * Format: date-time 
+       * Created at
+       * Format: date-time
+       *
        * @description The ISO 8601 timestamp that the member was created at
        */
       joinedAt: string;
       /**
-       * Is owner 
+       * Is owner
+       *
        * @default false
        */
       isOwner?: boolean;
@@ -1102,7 +1229,7 @@ export interface components {
     ServerMemberSummary: {
       user: components["schemas"]["UserSummary"];
       /** Role IDs */
-      roleIds: (number)[];
+      roleIds: number[];
     };
     /**
      * @example {
@@ -1114,36 +1241,42 @@ export interface components {
      */
     User: {
       /**
-       * User ID 
+       * User ID
+       *
        * @description The ID of the user
        */
       id: string;
       /**
-       * User type 
-       * @description The type of user. If this property is absent, it can assumed to be of type `user` 
+       * User type
+       *
+       * @description The type of user. If this property is absent, it can assumed to be of type `user`
        * @enum {string}
        */
       type?: "bot" | "user";
       /**
-       * User name 
+       * User name
+       *
        * @description The user's name
        */
       name: string;
       /**
-       * Avatar 
-       * Format: media-uri 
+       * Avatar
+       * Format: media-uri
+       *
        * @description The avatar image associated with the user
        */
       avatar?: string;
       /**
-       * Banner 
-       * Format: media-uri 
+       * Banner
+       * Format: media-uri
+       *
        * @description The banner image associated with the user
        */
       banner?: string;
       /**
-       * Created at 
-       * Format: date-time 
+       * Created at
+       * Format: date-time
+       *
        * @description The ISO 8601 timestamp that the user was created at
        */
       createdAt: string;
@@ -1157,24 +1290,28 @@ export interface components {
      */
     UserSummary: {
       /**
-       * User ID 
+       * User ID
+       *
        * @description The ID of the user
        */
       id: string;
       /**
-       * User type 
-       * @description The type of user. If this property is absent, it can assumed to be of type `user` 
+       * User type
+       *
+       * @description The type of user. If this property is absent, it can assumed to be of type `user`
        * @enum {string}
        */
       type?: "bot" | "user";
       /**
-       * User name 
+       * User name
+       *
        * @description The user's name
        */
       name: string;
       /**
-       * Avatar 
-       * Format: media-uri 
+       * Avatar
+       * Format: media-uri
+       *
        * @description The avatar image associated with the user
        */
       avatar?: string;
@@ -1194,18 +1331,21 @@ export interface components {
     ServerMemberBan: {
       user: components["schemas"]["UserSummary"];
       /**
-       * Reason 
+       * Reason
+       *
        * @description The reason for the ban as submitted by the banner
        */
       reason?: string;
       /**
-       * Created by 
+       * Created by
+       *
        * @description The ID of the user who created this server member ban
        */
       createdBy: string;
       /**
-       * Created at 
-       * Format: date-time 
+       * Created at
+       * Format: date-time
+       *
        * @description The ISO 8601 timestamp that the server member ban was created at
        */
       createdAt: string;
@@ -1224,69 +1364,80 @@ export interface components {
      */
     ServerChannel: {
       /**
-       * Channel ID 
-       * Format: uuid 
+       * Channel ID
+       * Format: uuid
+       *
        * @description The ID of the channel
        */
       id: string;
       /**
-       * @description The type of channel. This will determine what routes to use for creating content in a channel. For example, if this "chat", then one must use the routes for creating channel messages 
+       * @description The type of channel. This will determine what routes to use for creating content in a channel. For example, if this "chat", then one must use the routes for creating channel messages
        * @enum {string}
        */
-      type: "announcements" | "chat" | "calendar" | "forums" | "media" | "docs" | "voice" | "list" | "scheduling" | "stream";
+      type:
+        "announcements" | "calendar" | "chat" | "docs" | "forums" | "list" | "media" | "scheduling" | "stream" | "voice";
       /** @description The name of the channel */
       name: string;
       /** @description The topic of the channel */
       topic?: string;
       /**
-       * Created at 
-       * Format: date-time 
+       * Created at
+       * Format: date-time
+       *
        * @description The ISO 8601 timestamp that the channel was created at
        */
       createdAt: string;
       /**
-       * Created by 
+       * Created by
+       *
        * @description The ID of the user who created this channel
        */
       createdBy: string;
       /**
-       * Updated at 
-       * Format: date-time 
+       * Updated at
+       * Format: date-time
+       *
        * @description The ISO 8601 timestamp that the channel was updated at, if relevant
        */
       updatedAt?: string;
       /**
-       * Server ID 
+       * Server ID
+       *
        * @description The ID of the server
        */
       serverId: string;
       /**
-       * Channel ID 
-       * Format: uuid 
+       * Channel ID
+       * Format: uuid
+       *
        * @description ID of the parent channel or parent thread, if present. Only relevant for server channels
        */
       parentId?: string;
       /** @description Only relevant for server channels */
       categoryId?: number;
       /**
-       * Group ID 
+       * Group ID
+       *
        * @description The ID of the group
        */
       groupId: string;
       /**
-       * Is public 
-       * @description Whether the channel can be accessed from users who are not member of the server 
+       * Is public
+       *
+       * @description Whether the channel can be accessed from users who are not member of the server
        * @default false
        */
       isPublic?: boolean;
       /**
-       * Archived by 
+       * Archived by
+       *
        * @description The ID of the user who archived this channel
        */
       archivedBy?: string;
       /**
-       * Archived at 
-       * Format: date-time 
+       * Archived at
+       * Format: date-time
+       *
        * @description The ISO 8601 timestamp that the channel was archived at, if relevant
        */
       archivedAt?: string;
@@ -1306,67 +1457,80 @@ export interface components {
      */
     Server: {
       /**
-       * Server ID 
+       * Server ID
+       *
        * @description The ID of the server
        */
       id: string;
       /**
-       * Created by 
+       * Created by
+       *
        * @description The ID of the user who created this server
        */
       ownerId: string;
       /**
-       * Server type 
-       * @description The type of server designated from the server's settings page 
+       * Server type
+       *
+       * @description The type of server designated from the server's settings page
        * @enum {string}
        */
-      type?: "team" | "organization" | "community" | "clan" | "guild" | "friends" | "streaming" | "other";
+      type?:
+        "clan" | "community" | "friends" | "guild" | "organization" | "other" | "streaming" | "team";
       /**
-       * Server name 
+       * Server name
+       *
        * @description The name given to the server
        */
       name: string;
       /**
-       * Server URL 
+       * Server URL
+       *
        * @description The URL that the server can be accessible from. For example, a value of "Guilded-Official" means the server can be accessible from https://www.guilded.gg/Guilded-Official
        */
       url?: string;
       /**
-       * Description 
+       * Description
+       *
        * @description The description associated with the server
        */
       about?: string;
       /**
-       * Avatar 
-       * Format: media-uri 
+       * Avatar
+       * Format: media-uri
+       *
        * @description The avatar image associated with the server
        */
       avatar?: string;
       /**
-       * Banner 
-       * Format: media-uri 
+       * Banner
+       * Format: media-uri
+       *
        * @description The banner image associated with the server
        */
       banner?: string;
       /**
-       * Timezone 
+       * Timezone
+       *
        * @description The timezone associated with the server
        */
       timezone?: string;
       /**
-       * Is verified 
+       * Is verified
+       *
        * @description The verified status of the server
        */
       isVerified?: boolean;
       /**
-       * Channel ID 
-       * Format: uuid 
+       * Channel ID
+       * Format: uuid
+       *
        * @description The channel ID of the default channel of the server. This channel is defined as the first chat or voice channel in the left sidebar of a server in our UI. This channel is useful for sending welcome messages, though note that a bot may not have permissions to interact with this channel depending on how the server is configured.
        */
       defaultChannelId?: string;
       /**
-       * Created at 
-       * Format: date-time 
+       * Created at
+       * Format: date-time
+       *
        * @description The ISO 8601 timestamp that the server was created at
        */
       createdAt: string;
@@ -1383,52 +1547,61 @@ export interface components {
      */
     Webhook: {
       /**
-       * Webhook ID 
-       * Format: uuid 
+       * Webhook ID
+       * Format: uuid
+       *
        * @description The ID of the webhook
        */
       id: string;
       /**
-       * Name 
+       * Name
+       *
        * @description The name of the webhook
        */
       name: string;
       /**
-       * Avatar 
-       * Format: media-uri 
+       * Avatar
+       * Format: media-uri
+       *
        * @description The avatar image associated with the webhook
        */
       avatar?: string;
       /**
-       * Server ID 
+       * Server ID
+       *
        * @description The ID of the server
        */
       serverId: string;
       /**
-       * Channel ID 
-       * Format: uuid 
+       * Channel ID
+       * Format: uuid
+       *
        * @description The ID of the channel
        */
       channelId: string;
       /**
-       * Created at 
-       * Format: date-time 
+       * Created at
+       * Format: date-time
+       *
        * @description The ISO 8601 timestamp that the webhook was created at
        */
       createdAt: string;
       /**
-       * Created by 
+       * Created by
+       *
        * @description The ID of the user who created this webhook
        */
       createdBy: string;
       /**
-       * Deleted at 
-       * Format: date-time 
+       * Deleted at
+       * Format: date-time
+       *
        * @description The ISO 8601 timestamp that the webhook was deleted at
        */
       deletedAt?: string;
       /**
-       * Token 
+       * Token
+       *
        * @description The token of the webhook
        */
       token?: string;
@@ -1451,39 +1624,46 @@ export interface components {
      */
     CalendarEvent: {
       /**
-       * Calendar event ID 
+       * Calendar event ID
+       *
        * @description The ID of the calendar event
        */
       id: number;
       /**
-       * Server ID 
+       * Server ID
+       *
        * @description The ID of the server
        */
       serverId: string;
       /**
-       * Channel ID 
-       * Format: uuid 
+       * Channel ID
+       * Format: uuid
+       *
        * @description The ID of the channel
        */
       channelId: string;
       /**
-       * Name 
+       * Name
+       *
        * @description The name of the event
        */
       name: string;
       /**
-       * Description 
-       * Format: markdown 
+       * Description
+       * Format: markdown
+       *
        * @description The description of the event
        */
       description?: string;
       /**
-       * Location 
+       * Location
+       *
        * @description The location of the event
        */
       location?: string;
       /**
-       * Format: uri 
+       * Format: uri
+       *
        * @description A URL to associate with the event
        */
       url?: string;
@@ -1492,23 +1672,27 @@ export interface components {
       /** @description Is this event a repeating event */
       repeats?: boolean;
       /**
-       * Calendar event series ID 
-       * Format: uuid 
+       * Calendar event series ID
+       * Format: uuid
+       *
        * @description The ID of the calendar event series. Only shows if the event is repeating
        */
       seriesId?: string;
       /**
-       * Role IDs 
+       * Role IDs
+       *
        * @description The role IDs to restrict the event to
        */
-      roleIds?: (number)[];
+      roleIds?: number[];
       /**
-       * RSVP disabled 
+       * RSVP disabled
+       *
        * @description When disabled, users will not be able to RSVP to the event
        */
       rsvpDisabled?: boolean;
       /**
-       * Is all day 
+       * Is all day
+       *
        * @description Does the event last all day
        */
       isAllDay?: boolean;
@@ -1517,13 +1701,15 @@ export interface components {
       /** @description When `rsvpLimit` is set, users from the waitlist will be added as space becomes available in the event */
       autofillWaitlist?: boolean;
       /**
-       * Starts at 
-       * Format: date-time 
+       * Starts at
+       * Format: date-time
+       *
        * @description The ISO 8601 timestamp that the event starts at
        */
       startsAt: string;
       /**
-       * Duration 
+       * Duration
+       *
        * @description The duration of the event _**in minutes**_
        */
       duration?: number;
@@ -1531,25 +1717,29 @@ export interface components {
       isPrivate?: boolean;
       mentions?: components["schemas"]["Mentions"];
       /**
-       * Created at 
-       * Format: date-time 
+       * Created at
+       * Format: date-time
+       *
        * @description The ISO 8601 timestamp that the event was created at
        */
       createdAt: string;
       /**
-       * Created by 
+       * Created by
+       *
        * @description The ID of the user who created this event
        */
       createdBy: string;
       cancellation?: {
         /**
-         * Description 
-         * Format: markdown 
+         * Description
+         * Format: markdown
+         *
          * @description The description of event cancellation
          */
         description?: string;
         /**
-         * Created by 
+         * Created by
+         *
          * @description The ID of the user who created this event cancellation
          */
         createdBy: string;
@@ -1564,23 +1754,27 @@ export interface components {
      */
     Emote: {
       /**
-       * Emote ID 
+       * Emote ID
+       *
        * @description The ID of the emote
        */
       id: number;
       /**
-       * Name 
+       * Name
+       *
        * @description The name of the emote
        */
       name: string;
       /**
-       * Emote URL 
-       * Format: media-uri 
+       * Emote URL
+       * Format: media-uri
+       *
        * @description The URL of the emote image
        */
       url: string;
       /**
-       * Server ID 
+       * Server ID
+       *
        * @description The ID of the server the emote was created on
        */
       serverId?: string;
@@ -1598,51 +1792,61 @@ export interface components {
      */
     CalendarEventRsvp: {
       /**
-       * Calendar event ID 
+       * Calendar event ID
+       *
        * @description The ID of the calendar event
        */
       calendarEventId: number;
       /**
-       * Channel ID 
-       * Format: uuid 
+       * Channel ID
+       * Format: uuid
+       *
        * @description The ID of the channel
        */
       channelId: string;
       /**
-       * Server ID 
+       * Server ID
+       *
        * @description The ID of the server
        */
       serverId: string;
       /**
-       * User ID 
+       * User ID
+       *
        * @description The ID of the user
        */
       userId: string;
       /**
-       * Status 
-       * @description The status of the RSVP 
+       * Status
+       *
+       * @description The status of the RSVP
        * @enum {string}
        */
-      status: "going" | "maybe" | "declined" | "invited" | "waitlisted" | "not responded";
+      status:
+        "declined" | "going" | "invited" | "maybe" | "not responded" | "waitlisted";
       /**
-       * Created by 
+       * Created by
+       *
        * @description The ID of the user who created this RSVP
        */
       createdBy: string;
       /**
-       * Created at 
-       * Format: date-time 
+       * Created at
+       * Format: date-time
+       *
        * @description The ISO 8601 timestamp that the RSVP was created at
        */
       createdAt: string;
       /**
-       * Updated by 
+       * Updated by
+       *
        * @description The ID of the user who updated this RSVP
        */
       updatedBy?: string;
       /**
-       * Updated at 
-       * Format: date-time 
+       * Updated at
+       * Format: date-time
+       *
        * @description The ISO 8601 timestamp that the RSVP was updated at, if relevant
        */
       updatedAt?: string;
@@ -1659,41 +1863,48 @@ export interface components {
      */
     CalendarEventComment: {
       /**
-       * Calendar event comment ID 
+       * Calendar event comment ID
+       *
        * @description The ID of the calendar event comment
        */
       id: number;
       /**
-       * Content 
-       * Format: markdown 
+       * Content
+       * Format: markdown
+       *
        * @description The content of the calendar event comment
        */
       content: string;
       /**
-       * Created at 
-       * Format: date-time 
+       * Created at
+       * Format: date-time
+       *
        * @description The ISO 8601 timestamp that the calendar event comment was created at
        */
       createdAt: string;
       /**
-       * Updated at 
-       * Format: date-time 
+       * Updated at
+       * Format: date-time
+       *
        * @description The ISO 8601 timestamp that the calendar event comment was updated at, if relevant
        */
       updatedAt?: string;
       /**
-       * Calendar event ID 
+       * Calendar event ID
+       *
        * @description The ID of the calendar event
        */
       calendarEventId: number;
       /**
-       * Channel ID 
-       * Format: uuid 
+       * Channel ID
+       * Format: uuid
+       *
        * @description The ID of the channel
        */
       channelId: string;
       /**
-       * Created by 
+       * Created by
+       *
        * @description The ID of the user who created this calendar event comment (Note: If this event has `createdByWebhookId` present, this field will still be populated, but can be ignored. In this case, the value of this field will always be Ann6LewA)
        */
       createdBy: string;
@@ -1713,19 +1924,22 @@ export interface components {
      */
     CalendarEventReaction: {
       /**
-       * Channel ID 
-       * Format: uuid 
+       * Channel ID
+       * Format: uuid
+       *
        * @description The ID of the channel
        */
       channelId: string;
       /**
-       * User ID 
+       * User ID
+       *
        * @description The ID of the user who added the reaction
        */
       createdBy: string;
       emote: components["schemas"]["Emote"];
       /**
-       * Calendar event ID 
+       * Calendar event ID
+       *
        * @description The ID of the calendar event
        */
       calendarEventId: number;
@@ -1745,24 +1959,28 @@ export interface components {
      */
     CalendarEventCommentReaction: {
       /**
-       * Channel ID 
-       * Format: uuid 
+       * Channel ID
+       * Format: uuid
+       *
        * @description The ID of the channel
        */
       channelId: string;
       /**
-       * User ID 
+       * User ID
+       *
        * @description The ID of the user who added the reaction
        */
       createdBy: string;
       emote: components["schemas"]["Emote"];
       /**
-       * Calendar event ID 
+       * Calendar event ID
+       *
        * @description The ID of the calendar event
        */
       calendarEventId: number;
       /**
-       * Calendar event comment ID 
+       * Calendar event comment ID
+       *
        * @description The ID of the calendar event comment
        */
       calendarEventCommentId: number;
@@ -1776,19 +1994,22 @@ export interface components {
      */
     CalendarEventSeries: {
       /**
-       * Calendar event series ID 
-       * Format: uuid 
+       * Calendar event series ID
+       * Format: uuid
+       *
        * @description The ID of the calendar event series
        */
       id: string;
       /**
-       * Server ID 
+       * Server ID
+       *
        * @description The ID of the server
        */
       serverId: string;
       /**
-       * Channel ID 
-       * Format: uuid 
+       * Channel ID
+       * Format: uuid
+       *
        * @description The ID of the channel
        */
       channelId: string;
@@ -1806,41 +2027,48 @@ export interface components {
      */
     Announcement: {
       /**
-       * Announcement ID 
+       * Announcement ID
+       *
        * @description The ID of the announcement
        */
       id: string;
       /**
-       * Server ID 
+       * Server ID
+       *
        * @description The ID of the server
        */
       serverId: string;
       /**
-       * Channel ID 
-       * Format: uuid 
+       * Channel ID
+       * Format: uuid
+       *
        * @description The ID of the channel
        */
       channelId: string;
       /**
-       * Created at 
-       * Format: date-time 
+       * Created at
+       * Format: date-time
+       *
        * @description The ISO 8601 timestamp that the announcement was created at
        */
       createdAt: string;
       /**
-       * Created by 
+       * Created by
+       *
        * @description The ID of the user who created this announcement
        */
       createdBy: string;
       /**
-       * Content 
-       * Format: markdown 
+       * Content
+       * Format: markdown
+       *
        * @description The content of the announcement
        */
       content: string;
       mentions?: components["schemas"]["Mentions"];
       /**
-       * Title 
+       * Title
+       *
        * @description The title of the announcement
        */
       title: string;
@@ -1859,19 +2087,22 @@ export interface components {
      */
     AnnouncementReaction: {
       /**
-       * Channel ID 
-       * Format: uuid 
+       * Channel ID
+       * Format: uuid
+       *
        * @description The ID of the channel
        */
       channelId: string;
       /**
-       * User ID 
+       * User ID
+       *
        * @description The ID of the user who added the reaction
        */
       createdBy: string;
       emote: components["schemas"]["Emote"];
       /**
-       * Announcement ID 
+       * Announcement ID
+       *
        * @description The ID of the announcement
        */
       announcementId: string;
@@ -1888,41 +2119,48 @@ export interface components {
      */
     AnnouncementComment: {
       /**
-       * Announcement comment ID 
+       * Announcement comment ID
+       *
        * @description The ID of the announcement comment
        */
       id: number;
       /**
-       * Content 
-       * Format: markdown 
+       * Content
+       * Format: markdown
+       *
        * @description The content of the announcement comment
        */
       content: string;
       /**
-       * Created at 
-       * Format: date-time 
+       * Created at
+       * Format: date-time
+       *
        * @description The ISO 8601 timestamp that the announcement comment was created at
        */
       createdAt: string;
       /**
-       * Updated at 
-       * Format: date-time 
+       * Updated at
+       * Format: date-time
+       *
        * @description The ISO 8601 timestamp that the announcement comment was updated at, if relevant
        */
       updatedAt?: string;
       /**
-       * Created by 
+       * Created by
+       *
        * @description The ID of the user who created this announcement comment (Note: If this event has `createdByWebhookId` present, this field will still be populated, but can be ignored. In this case, the value of this field will always be Ann6LewA)
        */
       createdBy: string;
       /**
-       * Channel ID 
-       * Format: uuid 
+       * Channel ID
+       * Format: uuid
+       *
        * @description The ID of the channel
        */
       channelId: string;
       /**
-       * Announcement ID 
+       * Announcement ID
+       *
        * @description The ID of the announcement
        */
       announcementId: string;
@@ -1943,24 +2181,28 @@ export interface components {
      */
     AnnouncementCommentReaction: {
       /**
-       * Channel ID 
-       * Format: uuid 
+       * Channel ID
+       * Format: uuid
+       *
        * @description The ID of the channel
        */
       channelId: string;
       /**
-       * User ID 
+       * User ID
+       *
        * @description The ID of the user who added the reaction
        */
       createdBy: string;
       emote: components["schemas"]["Emote"];
       /**
-       * Announcement ID 
+       * Announcement ID
+       *
        * @description The ID of the announcement
        */
       announcementId: string;
       /**
-       * Announcement comment ID 
+       * Announcement comment ID
+       *
        * @description The ID of the announcement comment
        */
       announcementCommentId: number;
@@ -1975,8 +2217,9 @@ export interface components {
           /** @description The last message's ID that was sent to this bot. See [event replay](/docs/api/replay) for more details */
           lastMessageId: string;
           /**
-           * Bot ID 
-           * Format: uuid 
+           * Bot ID
+           * Format: uuid
+           *
            * @description The ID of the bot
            */
           botId: string;
@@ -1990,7 +2233,8 @@ export interface components {
         "application/json": {
           server: components["schemas"]["Server"];
           /**
-           * Created by 
+           * Created by
+           *
            * @description The ID of the user who created this server membership
            */
           createdBy: string;
@@ -2003,7 +2247,8 @@ export interface components {
         "application/json": {
           server: components["schemas"]["Server"];
           /**
-           * Deleted by 
+           * Deleted by
+           *
            * @description The ID of the user who deleted this server membership
            */
           deletedBy: string;
@@ -2014,7 +2259,8 @@ export interface components {
       content: {
         "application/json": {
           /**
-           * Server ID 
+           * Server ID
+           *
            * @description The ID of the server
            */
           serverId: string;
@@ -2026,7 +2272,8 @@ export interface components {
       content: {
         "application/json": {
           /**
-           * Server ID 
+           * Server ID
+           *
            * @description The ID of the server
            */
           serverId: string;
@@ -2038,31 +2285,36 @@ export interface components {
       content: {
         "application/json": {
           /**
-           * Server ID 
+           * Server ID
+           *
            * @description The ID of the server
            */
           serverId: string;
           message: {
             /**
-             * Message ID 
-             * Format: uuid 
+             * Message ID
+             * Format: uuid
+             *
              * @description The ID of the message
              */
             id: string;
             /**
-             * Server ID 
+             * Server ID
+             *
              * @description The ID of the server
              */
             serverId?: string;
             /**
-             * Channel ID 
-             * Format: uuid 
+             * Channel ID
+             * Format: uuid
+             *
              * @description The ID of the channel
              */
             channelId: string;
             /**
-             * Deleted at 
-             * Format: date-time 
+             * Deleted at
+             * Format: date-time
+             *
              * @description The ISO 8601 timestamp that the message was deleted at
              */
             deletedAt: string;
@@ -2076,7 +2328,8 @@ export interface components {
       content: {
         "application/json": {
           /**
-           * Server ID 
+           * Server ID
+           *
            * @description The ID of the server
            */
           serverId: string;
@@ -2088,22 +2341,26 @@ export interface components {
       content: {
         "application/json": {
           /**
-           * Server ID 
+           * Server ID
+           *
            * @description The ID of the server
            */
           serverId: string;
           /**
-           * User ID 
+           * User ID
+           *
            * @description The ID of the user
            */
           userId: string;
           /**
-           * Is kick? 
+           * Is kick?
+           *
            * @description If this member leaving was the result of a kick
            */
           isKick?: boolean;
           /**
-           * Is ban? 
+           * Is ban?
+           *
            * @description If this member leaving was the result of a ban
            */
           isBan?: boolean;
@@ -2114,7 +2371,8 @@ export interface components {
       content: {
         "application/json": {
           /**
-           * Server ID 
+           * Server ID
+           *
            * @description The ID of the server
            */
           serverId: string;
@@ -2126,7 +2384,8 @@ export interface components {
       content: {
         "application/json": {
           /**
-           * Server ID 
+           * Server ID
+           *
            * @description The ID of the server
            */
           serverId: string;
@@ -2138,18 +2397,21 @@ export interface components {
       content: {
         "application/json": {
           /**
-           * Server ID 
+           * Server ID
+           *
            * @description The ID of the server
            */
           serverId: string;
           userInfo: {
             /**
-             * User ID 
+             * User ID
+             *
              * @description The ID of the user
              */
             id: string;
             /**
-             * Nickname 
+             * Nickname
+             *
              * @description The nickname that was just updated for the user
              */
             nickname?: string | null;
@@ -2161,22 +2423,25 @@ export interface components {
       content: {
         "application/json": {
           /**
-           * Server ID 
+           * Server ID
+           *
            * @description The ID of the server
            */
           serverId: string;
-          memberRoleIds: ({
-              /**
-               * User ID 
-               * @description The ID of the member that had roles updated
-               */
-              userId: string;
-              /**
-               * Role IDs 
-               * @description The IDs of the roles that the member currently has _after_ this operation
-               */
-              roleIds: (number)[];
-            })[];
+          memberRoleIds: {
+            /**
+             * User ID
+             *
+             * @description The ID of the member that had roles updated
+             */
+            userId: string;
+            /**
+             * Role IDs
+             *
+             * @description The IDs of the roles that the member currently has _after_ this operation
+             */
+            roleIds: number[];
+          }[];
         };
       };
     };
@@ -2184,7 +2449,8 @@ export interface components {
       content: {
         "application/json": {
           /**
-           * Server ID 
+           * Server ID
+           *
            * @description The ID of the server
            */
           serverId: string;
@@ -2196,7 +2462,8 @@ export interface components {
       content: {
         "application/json": {
           /**
-           * Server ID 
+           * Server ID
+           *
            * @description The ID of the server
            */
           serverId: string;
@@ -2208,7 +2475,8 @@ export interface components {
       content: {
         "application/json": {
           /**
-           * Server ID 
+           * Server ID
+           *
            * @description The ID of the server
            */
           serverId: string;
@@ -2220,7 +2488,8 @@ export interface components {
       content: {
         "application/json": {
           /**
-           * Server ID 
+           * Server ID
+           *
            * @description The ID of the server
            */
           serverId: string;
@@ -2232,7 +2501,8 @@ export interface components {
       content: {
         "application/json": {
           /**
-           * Server ID 
+           * Server ID
+           *
            * @description The ID of the server
            */
           serverId: string;
@@ -2244,7 +2514,8 @@ export interface components {
       content: {
         "application/json": {
           /**
-           * Server ID 
+           * Server ID
+           *
            * @description The ID of the server
            */
           serverId: string;
@@ -2256,7 +2527,8 @@ export interface components {
       content: {
         "application/json": {
           /**
-           * Server ID 
+           * Server ID
+           *
            * @description The ID of the server
            */
           serverId: string;
@@ -2268,7 +2540,8 @@ export interface components {
       content: {
         "application/json": {
           /**
-           * Server ID 
+           * Server ID
+           *
            * @description The ID of the server
            */
           serverId: string;
@@ -2280,7 +2553,8 @@ export interface components {
       content: {
         "application/json": {
           /**
-           * Server ID 
+           * Server ID
+           *
            * @description The ID of the server
            */
           serverId: string;
@@ -2292,7 +2566,8 @@ export interface components {
       content: {
         "application/json": {
           /**
-           * Server ID 
+           * Server ID
+           *
            * @description The ID of the server
            */
           serverId: string;
@@ -2304,7 +2579,8 @@ export interface components {
       content: {
         "application/json": {
           /**
-           * Server ID 
+           * Server ID
+           *
            * @description The ID of the server
            */
           serverId: string;
@@ -2316,7 +2592,8 @@ export interface components {
       content: {
         "application/json": {
           /**
-           * Server ID 
+           * Server ID
+           *
            * @description The ID of the server
            */
           serverId: string;
@@ -2328,7 +2605,8 @@ export interface components {
       content: {
         "application/json": {
           /**
-           * Server ID 
+           * Server ID
+           *
            * @description The ID of the server
            */
           serverId: string;
@@ -2340,7 +2618,8 @@ export interface components {
       content: {
         "application/json": {
           /**
-           * Server ID 
+           * Server ID
+           *
            * @description The ID of the server
            */
           serverId: string;
@@ -2352,7 +2631,8 @@ export interface components {
       content: {
         "application/json": {
           /**
-           * Server ID 
+           * Server ID
+           *
            * @description The ID of the server
            */
           serverId: string;
@@ -2364,7 +2644,8 @@ export interface components {
       content: {
         "application/json": {
           /**
-           * Server ID 
+           * Server ID
+           *
            * @description The ID of the server
            */
           serverId: string;
@@ -2376,7 +2657,8 @@ export interface components {
       content: {
         "application/json": {
           /**
-           * Server ID 
+           * Server ID
+           *
            * @description The ID of the server
            */
           serverId: string;
@@ -2388,7 +2670,8 @@ export interface components {
       content: {
         "application/json": {
           /**
-           * Server ID 
+           * Server ID
+           *
            * @description The ID of the server
            */
           serverId: string;
@@ -2400,7 +2683,8 @@ export interface components {
       content: {
         "application/json": {
           /**
-           * Server ID 
+           * Server ID
+           *
            * @description The ID of the server
            */
           serverId: string;
@@ -2412,7 +2696,8 @@ export interface components {
       content: {
         "application/json": {
           /**
-           * Server ID 
+           * Server ID
+           *
            * @description The ID of the server
            */
           serverId: string;
@@ -2424,7 +2709,8 @@ export interface components {
       content: {
         "application/json": {
           /**
-           * Server ID 
+           * Server ID
+           *
            * @description The ID of the server
            */
           serverId: string;
@@ -2436,7 +2722,8 @@ export interface components {
       content: {
         "application/json": {
           /**
-           * Server ID 
+           * Server ID
+           *
            * @description The ID of the server
            */
           serverId: string;
@@ -2448,7 +2735,8 @@ export interface components {
       content: {
         "application/json": {
           /**
-           * Server ID 
+           * Server ID
+           *
            * @description The ID of the server
            */
           serverId: string;
@@ -2460,7 +2748,8 @@ export interface components {
       content: {
         "application/json": {
           /**
-           * Server ID 
+           * Server ID
+           *
            * @description The ID of the server
            */
           serverId: string;
@@ -2472,7 +2761,8 @@ export interface components {
       content: {
         "application/json": {
           /**
-           * Server ID 
+           * Server ID
+           *
            * @description The ID of the server
            */
           serverId: string;
@@ -2484,7 +2774,8 @@ export interface components {
       content: {
         "application/json": {
           /**
-           * Server ID 
+           * Server ID
+           *
            * @description The ID of the server
            */
           serverId: string;
@@ -2496,7 +2787,8 @@ export interface components {
       content: {
         "application/json": {
           /**
-           * Server ID 
+           * Server ID
+           *
            * @description The ID of the server
            */
           serverId: string;
@@ -2508,7 +2800,8 @@ export interface components {
       content: {
         "application/json": {
           /**
-           * Server ID 
+           * Server ID
+           *
            * @description The ID of the server
            */
           serverId: string;
@@ -2520,7 +2813,8 @@ export interface components {
       content: {
         "application/json": {
           /**
-           * Server ID 
+           * Server ID
+           *
            * @description The ID of the server
            */
           serverId: string;
@@ -2532,7 +2826,8 @@ export interface components {
       content: {
         "application/json": {
           /**
-           * Server ID 
+           * Server ID
+           *
            * @description The ID of the server
            */
           serverId: string;
@@ -2544,11 +2839,12 @@ export interface components {
       content: {
         "application/json": {
           /**
-           * Server ID 
+           * Server ID
+           *
            * @description The ID of the server
            */
           serverId: string;
-          calendarEventRsvps: (components["schemas"]["CalendarEventRsvp"])[];
+          calendarEventRsvps: components["schemas"]["CalendarEventRsvp"][];
         };
       };
     };
@@ -2556,7 +2852,8 @@ export interface components {
       content: {
         "application/json": {
           /**
-           * Server ID 
+           * Server ID
+           *
            * @description The ID of the server
            */
           serverId: string;
@@ -2568,7 +2865,8 @@ export interface components {
       content: {
         "application/json": {
           /**
-           * Server ID 
+           * Server ID
+           *
            * @description The ID of the server
            */
           serverId: string;
@@ -2580,7 +2878,8 @@ export interface components {
       content: {
         "application/json": {
           /**
-           * Server ID 
+           * Server ID
+           *
            * @description The ID of the server
            */
           serverId: string;
@@ -2592,7 +2891,8 @@ export interface components {
       content: {
         "application/json": {
           /**
-           * Server ID 
+           * Server ID
+           *
            * @description The ID of the server
            */
           serverId: string;
@@ -2604,7 +2904,8 @@ export interface components {
       content: {
         "application/json": {
           /**
-           * Server ID 
+           * Server ID
+           *
            * @description The ID of the server
            */
           serverId: string;
@@ -2616,7 +2917,8 @@ export interface components {
       content: {
         "application/json": {
           /**
-           * Server ID 
+           * Server ID
+           *
            * @description The ID of the server
            */
           serverId: string;
@@ -2628,7 +2930,8 @@ export interface components {
       content: {
         "application/json": {
           /**
-           * Server ID 
+           * Server ID
+           *
            * @description The ID of the server
            */
           serverId: string;
@@ -2640,12 +2943,14 @@ export interface components {
       content: {
         "application/json": {
           /**
-           * Server ID 
+           * Server ID
+           *
            * @description The ID of the server
            */
           serverId: string;
           /**
-           * Deleted by 
+           * Deleted by
+           *
            * @description The ID of the user who deleted this reaction
            */
           deletedBy: string;
@@ -2657,24 +2962,28 @@ export interface components {
       content: {
         "application/json": {
           /**
-           * Server ID 
+           * Server ID
+           *
            * @description The ID of the server
            */
           serverId: string;
           /**
-           * Channel ID 
-           * Format: uuid 
+           * Channel ID
+           * Format: uuid
+           *
            * @description The ID of the channel
            */
           channelId: string;
           /**
-           * Message ID 
-           * Format: uuid 
+           * Message ID
+           * Format: uuid
+           *
            * @description The ID of the message
            */
           messageId: string;
           /**
-           * Deleted by 
+           * Deleted by
+           *
            * @description The ID of the user who deleted this reaction
            */
           deletedBy: string;
@@ -2689,7 +2998,8 @@ export interface components {
       content: {
         "application/json": {
           /**
-           * Server ID 
+           * Server ID
+           *
            * @description The ID of the server
            */
           serverId: string;
@@ -2701,7 +3011,8 @@ export interface components {
       content: {
         "application/json": {
           /**
-           * Server ID 
+           * Server ID
+           *
            * @description The ID of the server
            */
           serverId: string;
@@ -2713,7 +3024,8 @@ export interface components {
       content: {
         "application/json": {
           /**
-           * Server ID 
+           * Server ID
+           *
            * @description The ID of the server
            */
           serverId: string;
@@ -2725,7 +3037,8 @@ export interface components {
       content: {
         "application/json": {
           /**
-           * Server ID 
+           * Server ID
+           *
            * @description The ID of the server
            */
           serverId: string;
@@ -2737,7 +3050,8 @@ export interface components {
       content: {
         "application/json": {
           /**
-           * Server ID 
+           * Server ID
+           *
            * @description The ID of the server
            */
           serverId: string;
@@ -2749,7 +3063,8 @@ export interface components {
       content: {
         "application/json": {
           /**
-           * Server ID 
+           * Server ID
+           *
            * @description The ID of the server
            */
           serverId: string;
@@ -2761,7 +3076,8 @@ export interface components {
       content: {
         "application/json": {
           /**
-           * Server ID 
+           * Server ID
+           *
            * @description The ID of the server
            */
           serverId: string;
@@ -2773,7 +3089,8 @@ export interface components {
       content: {
         "application/json": {
           /**
-           * Server ID 
+           * Server ID
+           *
            * @description The ID of the server
            */
           serverId: string;
@@ -2785,7 +3102,8 @@ export interface components {
       content: {
         "application/json": {
           /**
-           * Server ID 
+           * Server ID
+           *
            * @description The ID of the server
            */
           serverId: string;
@@ -2797,7 +3115,8 @@ export interface components {
       content: {
         "application/json": {
           /**
-           * Server ID 
+           * Server ID
+           *
            * @description The ID of the server
            */
           serverId: string;
@@ -2809,7 +3128,8 @@ export interface components {
       content: {
         "application/json": {
           /**
-           * Server ID 
+           * Server ID
+           *
            * @description The ID of the server
            */
           serverId: string;
@@ -2821,7 +3141,8 @@ export interface components {
       content: {
         "application/json": {
           /**
-           * Server ID 
+           * Server ID
+           *
            * @description The ID of the server
            */
           serverId: string;
@@ -2833,7 +3154,8 @@ export interface components {
       content: {
         "application/json": {
           /**
-           * Server ID 
+           * Server ID
+           *
            * @description The ID of the server
            */
           serverId: string;
@@ -2845,13 +3167,15 @@ export interface components {
       content: {
         "application/json": {
           /**
-           * Server ID 
+           * Server ID
+           *
            * @description The ID of the server
            */
           serverId: string;
           calendarEventSeries: components["schemas"]["CalendarEventSeries"];
           /**
-           * Calendar event ID 
+           * Calendar event ID
+           *
            * @description The calendar event updates started at
            */
           calendarEventId?: number;
@@ -2862,13 +3186,15 @@ export interface components {
       content: {
         "application/json": {
           /**
-           * Server ID 
+           * Server ID
+           *
            * @description The ID of the server
            */
           serverId: string;
           calendarEventSeries: components["schemas"]["CalendarEventSeries"];
           /**
-           * Calendar event ID 
+           * Calendar event ID
+           *
            * @description The calendar event deletions started at
            */
           calendarEventId?: number;
@@ -2879,7 +3205,8 @@ export interface components {
       content: {
         "application/json": {
           /**
-           * Server ID 
+           * Server ID
+           *
            * @description The ID of the server
            */
           serverId: string;
@@ -2891,7 +3218,8 @@ export interface components {
       content: {
         "application/json": {
           /**
-           * Server ID 
+           * Server ID
+           *
            * @description The ID of the server
            */
           serverId: string;
@@ -2903,7 +3231,8 @@ export interface components {
       content: {
         "application/json": {
           /**
-           * Server ID 
+           * Server ID
+           *
            * @description The ID of the server
            */
           serverId: string;
@@ -2915,7 +3244,8 @@ export interface components {
       content: {
         "application/json": {
           /**
-           * Server ID 
+           * Server ID
+           *
            * @description The ID of the server
            */
           serverId: string;
@@ -2927,7 +3257,8 @@ export interface components {
       content: {
         "application/json": {
           /**
-           * Server ID 
+           * Server ID
+           *
            * @description The ID of the server
            */
           serverId: string;
@@ -2939,7 +3270,8 @@ export interface components {
       content: {
         "application/json": {
           /**
-           * Server ID 
+           * Server ID
+           *
            * @description The ID of the server
            */
           serverId: string;
@@ -2951,7 +3283,8 @@ export interface components {
       content: {
         "application/json": {
           /**
-           * Server ID 
+           * Server ID
+           *
            * @description The ID of the server
            */
           serverId: string;
@@ -2963,7 +3296,8 @@ export interface components {
       content: {
         "application/json": {
           /**
-           * Server ID 
+           * Server ID
+           *
            * @description The ID of the server
            */
           serverId: string;
@@ -2975,7 +3309,8 @@ export interface components {
       content: {
         "application/json": {
           /**
-           * Server ID 
+           * Server ID
+           *
            * @description The ID of the server
            */
           serverId: string;
@@ -2987,7 +3322,8 @@ export interface components {
       content: {
         "application/json": {
           /**
-           * Server ID 
+           * Server ID
+           *
            * @description The ID of the server
            */
           serverId: string;
