+++ packages/api/lib/rest/errors/PermissionsError.ts
@@ -2,7 +2,11 @@ import type { RequestOptions, ResponseDetails } from "../RestManager";
 import { GuildedAPIError } from "./GuildedAPIError";
 
 export class PermissionsError extends GuildedAPIError {
-    public constructor(msg: string, request: RequestOptions, response: ResponseDetails) {
-        super(msg, request, response);
-    }
+  public constructor(
+    msg: string,
+    request: RequestOptions,
+    response: ResponseDetails
+  ) {
+    super(msg, request, response);
+  }
 }
