+++ packages/api/tsconfig.json
@@ -1,0 +1,9 @@
+{
+  "extends": "../../tsconfig.json",
+  "compilerOptions": {
+    "outDir": "./dist",
+    "declarationDir": "./types"
+  },
+  "include": ["./lib/**/*.ts"],
+  "exclude": ["*.test.ts"]
+}
