--- packages/guilded-api-typings/README.md
@@ -1,24 +1,0 @@
-# `@guildedjs/guilded-api-typings`
-
-[![GitHub](https://img.shields.io/badge/License-MIT-yellow.svg)](https://github.com/zaida04/guilded.js/blob/main/LICENSE)
-[![npm](https://img.shields.io/npm/v/@guildedjs/guilded-api-typings?color=crimson&logo=npm)](https://www.npmjs.com/package/@guildedjs/guilded-api-typings)
-[![CI workflows](https://github.com/zaida04/guilded.js/actions/workflows/ci.yml/badge.svg)](https://github.com/zaida04/guilded.js/actions/workflows/ci.yml)
-
-TypeScript Typings for the Guilded.gg API.
-
-> ⚠️ If you intend to use `guilded.js` you won't need to install this, as it's already a dependency of that package.
-
-## Installation
-
-You can install this package from [NPM](https://www.npmjs.com/package/@guildedjs/guilded-api-typings)
-
-- `npm install @guildedjs/guilded-api-typings`
-- `yarn add @guildedjs/guilded-api-typings`
-
-## Contributing
-
-Please see the main [README.md](https://github.com/zaida04/guilded.js) for info on how to contribute to this package or the other `@guildedjs` packages.
-
-## LICENSE
-
-Licensed under the [MIT License](https://github.com/zaida04/guilded.js/blob/main/LICENSE)
