--- packages/guilded-api-typings/lib/v1/structs/Channel.ts
@@ -1,1 +1,0 @@
-export type ChannelType = "announcements" | "calendar" | "chat" | "docs" | "forums" | "list" | "media" | "scheduling" | "stream" | "voice";
