--- packages/guilded-api-typings/package.json
@@ -1,66 +1,0 @@
-{
-  "name": "@guildedjs/guilded-api-typings",
-  "version": "4.0.0",
-  "description": "Typings for the guilded API",
-  "author": "Zaid \"Nico\" <contact@nico.engineer>",
-  "license": "MIT",
-  "types": "dist/index.d.ts",
-  "main": "dist/index.js",
-  "typedoc": {
-    "entryPoint": "lib/index.ts"
-  },
-  "scripts": {
-    "build": "tsc",
-    "build:typecheck": "tsc --noEmit",
-    "prepublishOnly": "rimraf dist/ && pnpm run build",
-    "generate:ws": "openapi-typescript \"https://www.guilded.gg/api/v1/socket-open-api-schema.json\" --output lib/v1/ws.ts",
-    "generate:rest": "openapi-typescript \"https://www.guilded.gg/api/v1/open-api-schema.json\" --output lib/v1/rest.ts"
-  },
-  "devDependencies": {
-    "openapi-typescript": "^6.2.4",
-    "typescript": "5.0.4"
-  },
-  "contributors": [
-    {
-      "name": "Zaid \"Nico\"",
-      "email": "contact@nico.engineer",
-      "url": "https://github.com/zaida04"
-    },
-    {
-      "name": "Skillz4Killz",
-      "email": "guildedjs@drskillz.33mail.com",
-      "url": "https://github.com/Skillz4Killz"
-    },
-    {
-      "name": "Uhuh \"Dylan\"",
-      "email": "dylan@panku.io",
-      "url": "https://github.com/uhuh"
-    },
-    {
-      "name": "DaStormer",
-      "email": "dastormer@stormdevelopmentz.xyz",
-      "url": "https://github.com/DaStormer"
-    }
-  ],
-  "keywords": [
-    "guilded",
-    "guildedjs",
-    "guilded.js",
-    "guilded-api",
-    "guilded-api-wrapper",
-    "guilded-wrapper",
-    "guilded-api-typings"
-  ],
-  "files": [
-    "dist"
-  ],
-  "homepage": "https://github.com/zaida04/guilded.js/tree/main/packages/guilded-api-typings#readme",
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/zaida04/guilded.js.git"
-  },
-  "bugs": {
-    "url": "https://github.com/zaida04/guilded.js/issues"
-  },
-  "gitHead": "eee38a19e0bfa812d7136cc78a6bc99e0b402b0c"
-}
