--- packages/guilded-api-typings/tsconfig.json
@@ -1,9 +1,0 @@
-{
-    "extends": "../../tsconfig.json",
-    "compilerOptions": {
-        "outDir": "./dist",
-        "target": "ES2020",
-        "removeComments": false
-    },
-    "include": ["./lib/**/*.ts"]
-}
