+++ packages/guilded.js/package.json
@@ -15,14 +15,12 @@
     "prepublishOnly": "rimraf dist/ && rimraf types/ && pnpm run build"
   },
   "devDependencies": {
-    "@guildedjs/guilded-api-typings": "workspace:*",
     "@types/node": "18.16.12",
     "typescript": "5.0.4"
   },
   "dependencies": {
     "@discordjs/collection": "^1.5.1",
-    "@guildedjs/rest": "workspace:*",
-    "@guildedjs/ws": "workspace:*",
+    "@guildedjs/api": "workspace:*",
     "typed-emitter": "2.1.0"
   },
   "contributors": [
