--- packages/ws/README.md
@@ -1,24 +1,0 @@
-# `@guildedjs/ws`
-
-[![GitHub](https://img.shields.io/badge/License-MIT-yellow.svg)](https://github.com/zaida04/guilded.js/blob/main/LICENSE)
-[![npm](https://img.shields.io/npm/v/@guildedjs/ws?color=crimson&logo=npm)](https://www.npmjs.com/package/@guildedjs/ws)
-[![CI workflows](https://github.com/zaida04/guilded.js/actions/workflows/ci.yml/badge.svg)](https://github.com/zaida04/guilded.js/actions/workflows/ci.yml)
-
-Utility for connecting to the Guilded.gg WebSocket gateway.
-
-> ⚠️ If you intend to use `guilded.js` you won't need to install this, as it's already a dependency of that package.
-
-## Installation
-
-You can install this package from [NPM](https://www.npmjs.com/package/@guildedjs/ws)
-
-- `npm install @guildedjs/ws`
-- `yarn add @guildedjs/ws`
-
-## Contributing
-
-Please see the main [README.md](https://github.com/zaida04/guilded.js) for info on how to contribute to this package or the other `@guildedjs` packages.
-
-## LICENSE
-
-Licensed under the [MIT License](https://github.com/zaida04/guilded.js/blob/main/LICENSE)
