--- packages/ws/lib/index.ts
@@ -1,1 +1,0 @@
-export * from "./WebSocketManager";
