--- packages/ws/package.json
@@ -1,74 +1,0 @@
-{
-  "name": "@guildedjs/ws",
-  "version": "1.7.1",
-  "description": "Module of guilded.js that handles WS connections",
-  "author": "Zaid \"Nico\" <contact@nico.engineer>",
-  "license": "MIT",
-  "types": "types/index.d.ts",
-  "main": "dist/index.js",
-  "typedoc": {
-    "entryPoint": "lib/index.ts"
-  },
-  "scripts": {
-    "build": "tsc && gen-esm-wrapper . ./dist/index.mjs",
-    "build:typecheck": "tsc --noEmit",
-    "prepublishOnly": "rimraf dist/ && rimraf types/ && pnpm run build"
-  },
-  "devDependencies": {
-    "typescript": "5.0.4"
-  },
-  "dependencies": {
-    "@guildedjs/guilded-api-typings": "workspace:*",
-    "@types/ws": "8.5.4",
-    "typed-emitter": "2.1.0",
-    "ws": "8.13.0"
-  },
-  "contributors": [
-    {
-      "name": "Zaid \"Nico\"",
-      "email": "contact@nico.engineer",
-      "url": "https://github.com/zaida04"
-    },
-    {
-      "name": "Skillz4Killz",
-      "email": "guildedjs@drskillz.33mail.com",
-      "url": "https://github.com/Skillz4Killz"
-    },
-    {
-      "name": "Uhuh \"Dylan\"",
-      "email": "dylan@panku.io",
-      "url": "https://github.com/uhuh"
-    },
-    {
-      "name": "DaStormer",
-      "email": "dastormer@stormdevelopmentz.xyz",
-      "url": "https://github.com/DaStormer"
-    }
-  ],
-  "exports": {
-    ".": {
-      "require": "./dist/index.js",
-      "import": "./dist/index.mjs"
-    },
-    "./": "./"
-  },
-  "keywords": [
-    "guilded",
-    "guildedjs",
-    "guilded.js",
-    "guilded-api"
-  ],
-  "files": [
-    "dist",
-    "types"
-  ],
-  "homepage": "https://github.com/zaida04/guilded.js/tree/main/packages/rest#readme",
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/zaida04/guilded.js.git"
-  },
-  "bugs": {
-    "url": "https://github.com/zaida04/guilded.js/issues"
-  },
-  "gitHead": "eee38a19e0bfa812d7136cc78a6bc99e0b402b0c"
-}
