--- packages/ws/tsconfig.json
@@ -1,9 +1,0 @@
-{
-    "extends": "../../tsconfig.json",
-    "compilerOptions": {
-        "outDir": "./dist",
-        "declarationDir": "./types"
-    },
-    "include": ["./lib/**/*.ts"],
-    "exclude": ["*.test.ts"]
-}
